Organized e2e environment managers into service-managers folder
- Moved managers (Ghost, MySQL, Portal, Tinybird) into service-managers/ subfolder for better code organization
- Reason for the update is since it aligns better with Docker Compose terminology
